agent/requirements.txt
# === Framework web ===
fastapi==0.119.1
uvicorn==0.38.0
starlette==0.48.0
sse-starlette==3.0.3
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.12.0
# === LangChain & LangGraph ===
langchain==1.0.7
langchain-openai==1.0.3
langchain-mistralai==1.0.1
langchain-google-genai==3.1.0
langchain-anthropic==1.1.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-classic==1.0.0
langchain-postgres==0.0.16
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-checkpoint-postgres==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langmem==0.0.30
# === LLM Providers & Embeddings ===
openai==2.8.1
anthropic==0.74.0
tiktoken==0.12.0
# === Base de donnees ===
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg2-binary==2.9.11
psycopg-pool==3.2.7
asyncpg==0.30.0
SQLAlchemy==2.0.41
# === Vectorstore ===
qdrant-client>=1.13.0
fastembed>=0.6.0
# === Observabilite ===
langfuse==3.10.0
langsmith==0.4.43
sentry-sdk==2.45.0
# === Utilitaires ===
python-dotenv==1.2.1
requests==2.32.5
requests-toolbelt==1.0.0
orjson==3.11.4
tenacity==9.1.2
python-dateutil==2.9.0.post0
